Move proxy service to signature package, split keypair service
