+++ src/middleware/packages/activitypub/index.js
@@ -11,7 +11,6 @@ module.exports = {
   InboxService: require('./services/inbox'),
   ObjectService: require('./services/object'),
   OutboxService: require('./services/outbox'),
-  ProxyService: require('./services/proxy'),
   RegistryService: require('./services/registry'),
   RelayService: require('./services/relay'),
   // Other services
